requirements.txt
dash
pandas
dash-auth
watchdog>=2.3
dash-bootstrap-components
dash-table
pillow>=10.0.0
plotly
pyzbar